requirements.txt
cs_gamestate==0.0.3
websocket_client==1.8.0
